// src/client/structures/enums.ts
export enum Cities {
	BRISBANE = 'r7hgdm',
	SYDNEY = 'r3gx2s',
	CANBERRA = 'r3dp5h',
	HOBART = 'r22u08',
	PERTH = 'qd66hr',
	DARWIN = 'qvv117',
	ADELAIDE = 'r1f93c',
	MELBOURNE = 'r1r0fu'
}

export enum RadarRegions {
	// Sydney region radars
	SYDNEY_TERREY_HILLS = 'IDR713',
	SYDNEY_WOLLONGONG = 'IDR712',
	
	// Melbourne region radars
	MELBOURNE = 'IDR023',
	
	// Brisbane region radars
	BRISBANE_MARBURG = 'IDR663',
	
	// Adelaide region radars
	ADELAIDE_BUCKLAND_PARK = 'IDR643',
	ADELAIDE_WOOMERA = 'IDR333',
	
	// Perth region radars
	PERTH = 'IDR703',
	
	// Darwin region radars
	DARWIN_BERRIMAH = 'IDR633',
	
	// Canberra region radars
	CANBERRA_CAPTAINS_FLAT = 'IDR403',
	
	// Hobart region radars
	HOBART = 'IDR423',
	
	// Other major radars
	NEWCASTLE = 'IDR033',
	NEWDEGATE = 'IDR313',
	CAIRNS = 'IDR303',
	TOWNSVILLE = 'IDR733',
	GLADSTONE = 'IDR363',
	GRAFTON = 'IDR043',
	PORT_HEDLAND = 'IDR313',
	BROOME = 'IDR293',
	ALICE_SPRINGS = 'IDR393'
}